refactor: eliminate settings duplication with shared DisplaySettingsList
Extract shared DisplaySettingsList component to handle group management
logic (getCleanGroups, handleGroupAction, handleUngroup), drag-sort, and
column rendering — previously duplicated ~430 lines across StandingsSettings
and RelativeSettings.

Add BaseDisplayConfig interface to widgetConfigs.ts so StandingsConfig and
RelativeConfig share their ~10 common fields via inheritance rather than
repetition. Each widget config now only declares its unique fields.
